+++ src/aitg_host/gens/summary_generator.py
@@ -1,26 +1,6 @@
 import torch
-from transformers import AutoTokenizer, AutoModelForSeq2SeqLM, logging
-from aitg_host.textgen.base_generator import BaseGenerator
+from aitg_host.gens.base import BaseGenerator
 from types import SimpleNamespace
-import re
-
-
-class SummarizerAI:
-    def __init__(self, model_folder: str, to_device, verbose=False):
-        self.model_folder = model_folder
-        self.device = to_device
-
-        if not verbose:
-            logging.set_verbosity_warning()
-
-        # load model
-        self.tokenizer = AutoTokenizer.from_pretrained(
-            model_folder, local_files_only=True
-        )
-        self.model = AutoModelForSeq2SeqLM.from_pretrained(
-            model_folder, local_files_only=True
-        ).to(to_device)
-
 
 class SummaryGenerator(BaseGenerator):
     def __init__(self, ai):
